Correctly implement support for changing CX mode on 2d raster.
Also, draw transparent areas when in pipe mode.

In pipe mode, parts of the plot area won't yet have data.  Draw these
as transparent instead of with the color.  This usually goes unnoticed
when the color mapped to 0 is black, but if the color mapped to 0 is not
black it looks like there is a bug.

We also start keeping track of absolute frame # because it will lay the
ground work for tracking of yaxis labels.