Fix for plugin canvas size remaining at 0
Plugins disappeared from the plot when `plot.checkresize()` was called.
Specifically, when the plot's parent div's height and width went to 0
(`display: none`), and `checkresize()` was called, and then the plot's
height and width returned to `display: block` with `checkresize()` called
again, the plugin canvas remained at height and width = 0.

The fix is re-adding a check to see whether plugin canvas size is
equal to plot canvas size, and if not, setting it.