Fix handling of pipe-raster zmin/zmax and colormap.
When zmin, zmax, and colormap are changed the image is invalidated,
but the logic in draw() was configured to not recreate the image
if the rendering was a pipe.  In hindsight there does not seem
to be a reason to restrict that logic to file mode only.

In addition, when the image was invalidated it also reset zmin/zmax,
but it should only do that if zmin and zmax are undefined.