Include in the dist folder a place to get the latest JS.
When building JSBINs or other web examples we need to be able to easily
point to the latest version.  Given that SigPlot will be moving away
from Bower the 'dist' folder on the primary repository will be removed.

Ideally we would have the output of travis-ci (or equivalent)
automatically build and publish build products somewhere, either
github or some JavaScript CDN.