fix: significant changes to p-cut implementation
The original p-cut implementation used separate canvases and changed the
styles of the parent canvas.  This could cause the plot to not render in
the desired position when placed on a page.

This implementation renders the p-cuts directly on the primary canvas
which avoids these issues, is more efficent, and requires less code.