chore: remove dependency of Array.prototype.remove shim
A shim was added to create a remove() method on Arrays, but this
can cause issues if someone attempts to use `for...in` on arrays.
Although `for...in` is *NOT* meant to iterate over arrays and doing
so is strongly discouraged by nearly all Javascript literature, it
is sometimes used for this purpose and adding methods to the Array
prototype will cause confusing behavior to those users.

Furthermore, it's probably considered poor practice for a library
(such as SigPlot) to alter things in the global namespace because
it could cause conflicts with the users application.