fix: handle x-cut/y-cut corner cases during zooms
First, refactor xCut/yCut logic to functions so the code is more
maintainable.

Second, when a cut is started on a zoom level handle that correctly
by creating a new zoom-stack and adjusting the axis as necessary.

Finally, when the cut is left go back to whereever we were when the
cut was entered.