Fix legend menu (#54)
* feat: add the ability to latch mouseup on widgets

If a widget is displayed due to left-button mouse then the same click
will trigger the widget action.  By setting Mx.mouseUpLatch you can
prevent this.

* fix: legend menu behaving incorrectly

The original code would show the menu when the mouse hovered and there
was a redraw.  It should only show when the mouse clicks on a legend
entry and then it should stay visible until an action is taken or the
menu is dismissed.