Fix unittests (#66)
* fix #63 by referencing a file that exists

* fix: #62

The original XPLOT would use "Unknown (U)" for the ylabel while
the xaxis would be "None (U)".  SigPlot now just uses "None (U)"
for both, so the unittest needs to be changed accordingly.

* fix: QUnit changed behavior of assert.expect

assert.expect(0) used to not itself count as an assertion; while
assert.expect(1) could have been used it's kinda awkward so instead
we just put an actual assertion.

* fix: scalar packed interactive test titles

The titles for the tests were copy pasted and inaccurate.

* chore: remove TBD from unittest description

* fix: zoomed x-cut unittest

The underlying file appears to have been changed, but the unittest
was not updated.